+++ src/Forms/src/Xamarin.Forms.Core/CornerRadiusTypeConverter.cs
@@ -15,38 +15,38 @@ namespace Xamarin.Forms
 				{ //Xaml
 					var cornerRadius = value.Split(',');
 					if (cornerRadius.Length == 4
-						&& double.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out double tl)
-						&& double.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out double tr)
-						&& double.TryParse(cornerRadius[2], NumberStyles.Number, CultureInfo.InvariantCulture, out double bl)
-						&& double.TryParse(cornerRadius[3], NumberStyles.Number, CultureInfo.InvariantCulture, out double br))
+						&& float.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out float tl)
+						&& float.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out float tr)
+						&& float.TryParse(cornerRadius[2], NumberStyles.Number, CultureInfo.InvariantCulture, out float bl)
+						&& float.TryParse(cornerRadius[3], NumberStyles.Number, CultureInfo.InvariantCulture, out float br))
 						return new CornerRadius(tl, tr, bl, br);
 					if (cornerRadius.Length > 1
 						&& cornerRadius.Length < 4
-						&& double.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out double l))
+						&& float.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out float l))
 						return new CornerRadius(l);
 				}
 				else if (value.Trim().Contains(" "))
 				{ //CSS
 					var cornerRadius = value.Split(' ');
 					if (cornerRadius.Length == 2
-						&& double.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out double t)
-						&& double.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out double b))
+						&& float.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out float t)
+						&& float.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out float b))
 						return new CornerRadius(t, b, b, t);
 					if (cornerRadius.Length == 3
-						&& double.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out double tl)
-						&& double.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out double trbl)
-						&& double.TryParse(cornerRadius[2], NumberStyles.Number, CultureInfo.InvariantCulture, out double br))
+						&& float.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out float tl)
+						&& float.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out float trbl)
+						&& float.TryParse(cornerRadius[2], NumberStyles.Number, CultureInfo.InvariantCulture, out float br))
 						return new CornerRadius(tl, trbl, trbl, br);
 					if (cornerRadius.Length == 4
-						&& double.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out tl)
-						&& double.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out double tr)
-						&& double.TryParse(cornerRadius[2], NumberStyles.Number, CultureInfo.InvariantCulture, out double bl)
-						&& double.TryParse(cornerRadius[3], NumberStyles.Number, CultureInfo.InvariantCulture, out br))
+						&& float.TryParse(cornerRadius[0], NumberStyles.Number, CultureInfo.InvariantCulture, out tl)
+						&& float.TryParse(cornerRadius[1], NumberStyles.Number, CultureInfo.InvariantCulture, out float tr)
+						&& float.TryParse(cornerRadius[2], NumberStyles.Number, CultureInfo.InvariantCulture, out float bl)
+						&& float.TryParse(cornerRadius[3], NumberStyles.Number, CultureInfo.InvariantCulture, out br))
 						return new CornerRadius(tl, tr, bl, br);
 				}
 				else
 				{ //single uniform CornerRadius
-					if (double.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out double l))
+					if (float.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out float l))
 						return new CornerRadius(l);
 				}
 			}
