+++ src/Forms/src/Xamarin.Forms.Core/Easing.cs
@@ -33,14 +33,14 @@ namespace Xamarin.Forms
 	{
 		public static readonly Easing Linear = new Easing(x => x);
 
-		public static readonly Easing SinOut = new Easing(x => Math.Sin(x * Math.PI * 0.5f));
-		public static readonly Easing SinIn = new Easing(x => 1.0f - Math.Cos(x * Math.PI * 0.5f));
-		public static readonly Easing SinInOut = new Easing(x => -Math.Cos(Math.PI * x) / 2.0f + 0.5f);
+		public static readonly Easing SinOut = new Easing(x => (float)Math.Sin(x * Math.PI * 0.5f));
+		public static readonly Easing SinIn = new Easing(x => 1.0f - (float)Math.Cos(x * Math.PI * 0.5f));
+		public static readonly Easing SinInOut = new Easing(x => -(float)Math.Cos(Math.PI * x) / 2.0f + 0.5f);
 
 		public static readonly Easing CubicIn = new Easing(x => x * x * x);
-		public static readonly Easing CubicOut = new Easing(x => Math.Pow(x - 1.0f, 3.0f) + 1.0f);
+		public static readonly Easing CubicOut = new Easing(x => (float)Math.Pow(x - 1.0f, 3.0f) + 1.0f);
 
-		public static readonly Easing CubicInOut = new Easing(x => x < 0.5f ? Math.Pow(x * 2.0f, 3.0f) / 2.0f : (Math.Pow((x - 1) * 2.0f, 3.0f) + 2.0f) / 2.0f);
+		public static readonly Easing CubicInOut = new Easing(x => x < 0.5f ? (float)Math.Pow(x * 2.0f, 3.0f) / 2.0f : (float)(Math.Pow((x - 1) * 2.0f, 3.0f) + 2.0f) / 2.0f);
 
 		public static readonly Easing BounceOut;
 		public static readonly Easing BounceIn;
@@ -48,7 +48,7 @@ namespace Xamarin.Forms
 		public static readonly Easing SpringIn = new Easing(x => x * x * ((1.70158f + 1) * x - 1.70158f));
 		public static readonly Easing SpringOut = new Easing(x => (x - 1) * (x - 1) * ((1.70158f + 1) * (x - 1) + 1.70158f) + 1);
 
-		readonly Func<double, double> _easingFunc;
+		readonly Func<float, float> _easingFunc;
 
 		static Easing()
 		{
@@ -78,7 +78,7 @@ namespace Xamarin.Forms
 			BounceIn = new Easing(p => 1.0f - BounceOut.Ease(1 - p));
 		}
 
-		public Easing(Func<double, double> easingFunc)
+		public Easing(Func<float, float> easingFunc)
 		{
 			if (easingFunc == null)
 				throw new ArgumentNullException("easingFunc");
@@ -86,12 +86,12 @@ namespace Xamarin.Forms
 			_easingFunc = easingFunc;
 		}
 
-		public double Ease(double v)
+		public float Ease(float v)
 		{
 			return _easingFunc(v);
 		}
 
-		public static implicit operator Easing(Func<double, double> func)
+		public static implicit operator Easing(Func<float, float> func)
 		{
 			return new Easing(func);
 		}
