+++ src/Forms/src/Xamarin.Forms.Core/IFontElement.cs
@@ -10,12 +10,12 @@ namespace Xamarin.Forms.Internals
 		string FontFamily { get; }
 
 		[TypeConverter(typeof(FontSizeConverter))]
-		double FontSize { get; }
+		float FontSize { get; }
 
 		//note to implementor: but implement these methods explicitly
 		void OnFontFamilyChanged(string oldValue, string newValue);
-		void OnFontSizeChanged(double oldValue, double newValue);
-		double FontSizeDefaultValueCreator();
+		void OnFontSizeChanged(float oldValue, float newValue);
+		float FontSizeDefaultValueCreator();
 		void OnFontAttributesChanged(FontAttributes oldValue, FontAttributes newValue);
 		void OnFontChanged(Font oldValue, Font newValue);
 	}
